add changes for 3 homework
