+++ simple-chat/src/index.css
@@ -182,123 +182,3 @@ input[type=submit] {
 
 
 
-/*  НАЧАЛО
-    Стили для страницы с чатами (для макета из ДЗ-3) */
-
-.container2 {
-    display: flex;
-    flex-direction: column;
-    min-height: 100vh;
-}
-
-/*  НАЧАЛО
-    Стили для блока header и его дочерних элементов */
-header {
-    display: flex;
-    align-items: center;
-    background-color: #8E24AA;
-    color: #fff;
-    font-size: calc(10px + 24 * (100vw / 1280));
-    padding: 2vh 0;
-}
-
-.menu {
-    margin-left: 2vw;
-    margin-right: 10vw;
-}
-
-.search {
-    margin-left: auto;
-    margin-right: 2vw;
-}
-
-.photo img{
-    border-radius: 50%;
-    width:8vh;
-}
-/*  КОНЕЦ
-    Стили для блока header и его дочерних элементов */
-
-/*  НАЧАЛО
-    Стили для блока chats и его дочерних элементов */
-
-.chats {
-    overflow: auto;
-    height: 100vh;
-}
-.chat {
-    display: flex;
-    align-items: stretch;
-    margin-top: 1vh;
-}
-
-.container2 .photo {
-    margin-left: 2vw;
-    margin-right: 10vw;
-    display: flex;
-    align-items: center;
-}
-
-.chat-info {
-    padding-top: 1vh;
-    border-bottom: 1px solid #d1d1d1;
-    flex: 1;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-}
-
-.name-text {
-    height: 100%;
-}
-.name {
-    /* padding-top: 1vh ; */
-    font-size: calc(14px + 24 * (100vw / 1280));
-}
-
-.text {
-    white-space: nowrap;
-    padding-top: 1vh;
-    color: grey;
-    font-size: calc(10px + 24 * (100vw / 1280));
-    overflow: hidden;
-    text-overflow: ellipsis;
-    min-height: 5vh;
-    max-width: 55vw;
-}
-
-.container2 .info {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-end;
-    height: 100%;
-    margin-right: 2vw;
-}
-
-.condition {
-    padding-top: 1vh;
-    color: #8E24AA;
-}
-/*  КОНЕЦ
-    Стили для блока chats и его дочерних элементов */
-
-/*  НАЧАЛО
-    Стили для блока create-chat и его дочерних элементов */
-.create-chat {
-    color: grey;
-    position: absolute;
-    bottom: 5vh;
-    right: 5vh;
-    background-color: #ffca00ba;
-    height: 8vh;
-    width: 8vh;
-    display: flex;
-    align-content: center;
-    align-items: center;
-    justify-content: center;
-    border-radius: 50%;
-    box-shadow: 0px 5px 7px 1px rgb(0 0 0 / 20%);
-}
-
-/*  Стили для страницы с чатами (для макета из ДЗ-3)
-    КОНЕЦ */
