+++ simple-chat/src/index.html
@@ -5,6 +5,7 @@
         <meta name="viewport" content="width=device-width, initial-scale=1.0">
         <title>Send a message</title>
         <link rel="stylesheet" href="./index.css">
+        <link rel="stylesheet" href="./chats.css">
         <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
     </head>
     <body>
@@ -88,5 +89,6 @@
             </div>
         </div>  
         <script type="text/javascript" src="./index.js"></script>
+        <script type="text/javascript" src="./chats.js"></script>
     </body>
 </html>
